Move proc_file_utils to //cuttlefish/process
Bug: b/533572315
