+++ base/cvd/cuttlefish/host/commands/cvd/instances/BUILD.bazel
@@ -106,7 +106,6 @@ cf_cc_library(
         "//cuttlefish/common/libs/utils:files",
         "//cuttlefish/common/libs/utils:gflags_xml_parser",
         "//cuttlefish/common/libs/utils:json",
-        "//cuttlefish/common/libs/utils:proc_file_utils",
         "//cuttlefish/common/libs/utils:signals",
         "//cuttlefish/common/libs/utils:users",
         "//cuttlefish/host/commands/cvd/cli:command_request",
@@ -128,6 +127,7 @@ cf_cc_library(
         "//cuttlefish/posix:symlink",
         "//cuttlefish/process:command",
         "//cuttlefish/process:managed_stdio",
+        "//cuttlefish/process:proc_file_utils",
         "//cuttlefish/result",
         "//libbase",
         "@abseil-cpp//absl/log",
@@ -150,7 +150,6 @@ cf_cc_library(
         "//cuttlefish/common/libs/utils:contains",
         "//cuttlefish/common/libs/utils:files",
         "//cuttlefish/common/libs/utils:json",
-        "//cuttlefish/common/libs/utils:proc_file_utils",
         "//cuttlefish/common/libs/utils:signals",
         "//cuttlefish/common/libs/utils:users",
         "//cuttlefish/host/commands/cvd/cli:command_request",
@@ -172,6 +171,7 @@ cf_cc_library(
         "//cuttlefish/posix:symlink",
         "//cuttlefish/process:command",
         "//cuttlefish/process:managed_stdio",
+        "//cuttlefish/process:proc_file_utils",
         "//cuttlefish/result",
         "//libbase",
         "@abseil-cpp//absl/log",
@@ -205,9 +205,9 @@ cf_cc_library(
     deps = [
         "//cuttlefish/common/libs/utils:contains",
         "//cuttlefish/common/libs/utils:files",
-        "//cuttlefish/common/libs/utils:proc_file_utils",
         "//cuttlefish/host/commands/cvd/utils:common",
         "//cuttlefish/host/libs/config:config_constants",
+        "//cuttlefish/process:proc_file_utils",
         "//cuttlefish/result",
         "//libbase",
         "@abseil-cpp//absl/log",
@@ -222,7 +222,6 @@ cf_cc_library(
     deps = [
         "//cuttlefish/common/libs/utils:contains",
         "//cuttlefish/common/libs/utils:files",
-        "//cuttlefish/common/libs/utils:proc_file_utils",
         "//cuttlefish/host/commands/cvd/instances:config_path",
         "//cuttlefish/host/commands/cvd/instances:run_cvd_proc_collector",
         "//cuttlefish/host/commands/cvd/utils",
@@ -230,6 +229,7 @@ cf_cc_library(
         "//cuttlefish/posix:strerror",
         "//cuttlefish/process:command",
         "//cuttlefish/process:managed_stdio",
+        "//cuttlefish/process:proc_file_utils",
         "//cuttlefish/result",
         "//libbase",
         "@abseil-cpp//absl/log",
