+++ base/cvd/cuttlefish/host/commands/cvd/instances/run_cvd_proc_collector.cpp
@@ -32,9 +32,9 @@
 
 #include "cuttlefish/common/libs/utils/contains.h"
 #include "cuttlefish/common/libs/utils/files.h"
-#include "cuttlefish/common/libs/utils/proc_file_utils.h"
 #include "cuttlefish/host/commands/cvd/utils/common.h"
 #include "cuttlefish/host/libs/config/config_constants.h"
+#include "cuttlefish/process/proc_file_utils.h"
 
 namespace cuttlefish {
 namespace {
